move redis config
