+++ src/main/java/uk/gov/ons/ctp/response/collection/exercise/config/RedisConfig.java
@@ -1,14 +1,6 @@
 package uk.gov.ons.ctp.response.collection.exercise.config;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.jedis.JedisClientConfiguration;
-import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
 
 @Configuration
-public class RedisConfig {
-
-}
+public class RedisConfig {}
